pyproject.toml
[tool.poetry]
name = "bh-testing"
version = "0.1.0"
description = "An Schwartzschild black hole image generator."
authors = ["pedropasa03"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.1.1"
pillow = "^10.4.0"
moderngl = "^5.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
